[NEW] moved personnality, reorder
