+++ static/personalities.json
